+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# <a href='https://github.com/mrjackwills/staticpi_backend/releases/tag/v0.0.17'>v0.0.17</a>
+### 2024-08-14
+
+### Chores
++ dependencies updated, [6cb671e3](https://github.com/mrjackwills/staticpi_backend/commit/6cb671e37854f5da3ff25b2d5650281e97fabd9d),
+
+### Features
++ improve healthcheck, should now be *correct*, [e89e39a2](https://github.com/mrjackwills/staticpi_backend/commit/e89e39a241f29cf51b68e16e7d280556859673b5),
+
+### Refactors
++ switch from `/dev/shm` to `/ramdrive`, [24160e54](https://github.com/mrjackwills/staticpi_backend/commit/24160e54da206261606c898368ba5ff796be557c),
+
 # <a href='https://github.com/mrjackwills/staticpi_backend/releases/tag/v0.0.16'>v0.0.16</a>
 ### 2024-08-08